Adds more info to right side bar, updates Game API, and adds colors!
I added the info for higher level spells to the right side bar to make
it more clear what is happening. While doing this I noticed that some of
the higher spell functions (such as focus) weren't fully implemented.
This was partially due to the poor Game API (a bunch of public
variables). So I updated the API to actually use functions and made
(almost) all the variables private. Finally I added some new colors to
make the side bar easier to read.